astro_engine/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
pyswisseph
waitress>=2.0.0
gunicorn>=20.0.0
pytz>=2021.0
python-dateutil>=2.8.0
redis>=4.0.0
Flask-Limiter>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.0
pybreaker>=1.4.0
pydantic>=2.0.0
Flask-Compress>=1.20
rq>=2.6.0
tenacity>=9.0.0
